Change the logic for Steam private libs to a list of libs to override
Currently the Steam private libs (Steam runtime heavy) is a moving target for
both compatibility and its contents. It has been common for Steam beta clients
to break in LSI with incompatible library changes and new additions that aren't
installed in many systems. By changing the logic, it should become more
resistant to changes in Steam runtime heavy. The runtime is updated frequently
and not meant to retain compatibility.

Based on proc maps for the current client, the only difference now is libSDL2,
which has recently shown to be the most likely library to use an unreleased
version.